Refactor training part of `engine` module. (#10029)
* Refactor topological part of Keras engine.

* Fix imports

* Fix merge mixup.

* Refactor training part of the Keras engine.

* Fix unit tests.
